device: Delete shared copy rules for sensor configuration
We have never had copy rules nor sensor files in device-specific trees.
These files are supplied by `prebuilt_etc` declarations in common (no
need for a copy rule, just depend on the file name which is a "prebuilt"
package) and generically included by the platform repository; Edo is no
exception and has them added in [1]. The (list of) included files is
identical between PDX203 and PDX206.

[1]: https://github.com/sonyxperiadev/device-sony-edo/pull/19