sensors: Configure device-specific orientation of LSM6DSM/AK0991
PDX203 and PDX206 have the LSM6DSM 6DoF IMU (Inertial Measurement Unit)
and AK0991 magnetometer mounted at a different orientation, resulting in
one of the two devices orienting itself wrongly.  That difference is
compensated for in aptly-named `_somc_product` overlay files, which are
moved into its corresponding device repo here to account for the
difference in the `".orient"` section.